requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=4.0.0
pytz>=2023.3
requests>=2.31.0
